# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get the current authenticated user (provisions on first call)
      responses:
        "200":
          description: The current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update the current user's profile or role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /teachers/{teacherId}/profile:
    get:
      operationId: getTeacherProfile
      tags: [users]
      summary: Get a teacher's public profile by id
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teacher profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by id
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /courses/{courseId}/materials:
    get:
      operationId: listCourseMaterials
      tags: [materials]
      summary: List course materials (teacher and enrolled students)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseMaterial"
    post:
      operationId: createCourseMaterial
      tags: [materials]
      summary: Post a course material (teacher only)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseMaterialInput"
      responses:
        "201":
          description: Created material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseMaterial"
  /courses/{courseId}/materials/unseen:
    get:
      operationId: getUnseenMaterialsCount
      tags: [materials]
      summary: Count materials posted since the current user last viewed the Materials tab
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Unseen materials count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnseenMaterialsCount"
  /courses/{courseId}/materials/seen:
    post:
      operationId: markMaterialsSeen
      tags: [materials]
      summary: Mark all course materials as seen for the current user
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "204":
          description: Marked as seen
  /materials/{materialId}:
    patch:
      operationId: updateCourseMaterial
      tags: [materials]
      summary: Update a course material (teacher only)
      parameters:
        - $ref: "#/components/parameters/MaterialId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseMaterialInput"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseMaterial"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteCourseMaterial
      tags: [materials]
      summary: Delete a course material (teacher only)
      parameters:
        - $ref: "#/components/parameters/MaterialId"
      responses:
        "204":
          description: Material deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: integer
...
    MaterialId:
      name: materialId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, teacher, dean, course_coordinator, unassigned]
        bio:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
        activeCourseCount:
          type: integer
...
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [student, teacher]
        bio:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
    TeacherProfile:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
    Course:
      type: object
      required: [id, title, teacherId, inviteCode, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        teacherId:
          type: string
        teacherName:
          type: ["string", "null"]
        inviteCode:
          type: string
        isActive:
          type: boolean
        studentCount:
          type: integer
        assignmentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    MaterialFile:
      type: object
      required: [path, name]
      properties:
        path:
          type: string
        name:
          type: string
        size:
          type: integer
    CourseMaterial:
      type: object
      required: [id, courseId, authorId, title, links, attachments, createdAt, updatedAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        authorId:
          type: string
        title:
          type: string
        body:
          type: ["string", "null"]
        links:
          type: array
          items:
            type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/MaterialFile"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UnseenMaterialsCount:
      type: object
      required: [count]
      properties:
        count:
          type: integer
    CourseMaterialInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
        links:
          type: array
          items:
            type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/MaterialFile"
...